moved bot test to new package and added some more comments to code
