+++ src/main/java/backend/bot/MediumBot.java
@@ -13,12 +13,14 @@ public class MediumBot extends BotBase {
         super(token);
     }
     public int makeMove(VierGewinnt vg) {
+
         // Check if there is a column where the bot can win
         for (int i = 1; i <= vg.getNumberOfColumns(); i++) {
             if (!vg.columnIsFull(i)) {
                 VierGewinnt tmpVg = vg.copy();
                 tmpVg.addToken(i, super.getMyToken());
                 if (tmpVg.checkForWin() == super.getMyToken().toGameStatus()) {
+                    // Add a token to the column
                     vg.addToken(i, super.getMyToken());
                     return i;
                 }
@@ -31,6 +33,7 @@ public class MediumBot extends BotBase {
                 VierGewinnt tmpVg = vg.copy();
                 tmpVg.addToken(i, super.getMyToken().other());
                 if (tmpVg.checkForWin() == super.getMyToken().other().toGameStatus()) {
+                    // Add a token to the column
                     vg.addToken(i, super.getMyToken());
                     return i;
                 }
@@ -43,6 +46,7 @@ public class MediumBot extends BotBase {
             column = (int) (Math.random() * vg.getNumberOfColumns()) + 1;
         } while (vg.columnIsFull(column));
 
+        // Add a token to the column
         vg.addToken(column, super.getMyToken());
 
         return column;
